refactor: moved JSONLd & Helmet from prodDetail to prod page, Helmet clean-up
