+++ src/components/ProductDetail/ProductDetail.js
@@ -3,7 +3,6 @@ import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import { inject, observer } from "mobx-react";
-import Helmet from "react-helmet";
 import track from "lib/tracking/track";
 import Breadcrumbs from "components/Breadcrumbs";
 import trackProductViewed from "lib/tracking/trackProductViewed";
@@ -57,7 +56,7 @@ class ProductDetail extends Component {
     }),
     theme: PropTypes.object,
     uiStore: PropTypes.object.isRequired
-  }
+  };
 
   componentDidMount() {
     const { product } = this.props;
@@ -110,10 +109,7 @@ class ProductDetail extends Component {
       addItemsToCart,
       currencyCode,
       product,
-      uiStore: {
-        pdpSelectedOptionId,
-        pdpSelectedVariantId
-      }
+      uiStore: { pdpSelectedOptionId, pdpSelectedVariantId }
     } = this.props;
 
     // Get selected variant or variant option
@@ -181,51 +177,6 @@ class ProductDetail extends Component {
     return productPrice;
   }
 
-  renderJSONLd() {
-    const { currencyCode, product, shop } = this.props;
-
-    const priceData = product.pricing[0];
-    const images = product.media.map((image) => image.URLs.original);
-
-    let productAvailability = "http://schema.org/InStock";
-    if (product.isLowQuantity) {
-      productAvailability = "http://schema.org/LimitedAvailability";
-    }
-    if (product.isBackorder && product.isSoldOut) {
-      productAvailability = "http://schema.org/PreOrder";
-    }
-    if (!product.isBackorder && product.isSoldOut) {
-      productAvailability = "http://schema.org/SoldOut";
-    }
-
-    // Recommended data from https://developers.google.com/search/docs/data-types/product
-    const productJSON = {
-      "@context": "http://schema.org/",
-      "@type": "Product",
-      "brand": product.vendor,
-      "description": product.description,
-      "image": images,
-      "name": product.title,
-      "sku": product.sku,
-      "offers": {
-        "@type": "Offer",
-        "priceCurrency": currencyCode,
-        "price": priceData.minPrice,
-        "availability": productAvailability,
-        "seller": {
-          "@type": "Organization",
-          "name": shop.name
-        }
-      }
-    };
-
-    return (
-      <script type="application/ld+json">
-        {JSON.stringify(productJSON)}
-      </script>
-    );
-  }
-
   render() {
     const {
       classes,
@@ -266,11 +217,6 @@ class ProductDetail extends Component {
 
     return (
       <Fragment>
-        <Helmet>
-          <title>{product.title}</title>
-          <meta name="description" content={product.description} />
-          {this.renderJSONLd()}
-        </Helmet>
         <Grid container spacing={theme.spacing.unit * 3}>
           <Grid item className={classes.breadcrumbGrid} xs={12}>
             <Breadcrumbs isPDP={true} tag={tag} tags={tags} product={product} />
@@ -285,10 +231,7 @@ class ProductDetail extends Component {
           </Grid>
 
           <Grid item xs={12} sm={6}>
-            <ProductDetailTitle
-              pageTitle={product.pageTitle}
-              title={product.title}
-            />
+            <ProductDetailTitle pageTitle={product.pageTitle} title={product.title} />
             <ProductDetailInfo
               priceRange={productPrice.displayPrice}
               description={product.description}
